Bug 694810 - Support the Opus codec for webrtc. r=jesup (mozilla (non-webrtc.org) parts)
Add support for the Opus audio codec to the webrtc library,
and use it the corresponding unit tests. Support is only
mono for now.

Since the neteq library only supports audio packets up
to 32 kHz, we run the opus codec at its native 48 kHz,
and also use that rate for the RTP timestamp clock, as
per the payload draft. However we resample the actual
audio data to/from 32 kHz in the WebRtcOpus_* wrapper
code so that webrtc internally sees data at its highest
supported rate.

Internal webrtc audio buffer sizes, which are generally
implicit, are doubled to make room for the maximum Opus
packet size of 120 ms. This isn't tested.

Thanks to Tim Terriberry, Jean-Marc Valin and others for
help developing this patch.